w2d5,forge里script使用

+++ w2d4-foundry/src/Bank.sol
@@ -1,11 +1,11 @@
-// SPDX-License-Identifier: MIT 
+// SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.13;
 
 contract Bank {
-    mapping(address => uint) public balanceOf;
+    mapping(address => uint256) public balanceOf;
 
-    event Deposit(address indexed user, uint amount);
+    event Deposit(address indexed user, uint256 amount);
 
     function depositETH() external payable {
         require(msg.value > 0, "Deposit amount must be greater than 0");
